small modifications

+++ v1/src/main/kotlin/base/Vertex.kt
@@ -1,10 +1,10 @@
 package base
 
-data class Vertex(val x: Float,val y: Float){
+data class Vertex(val x: Double,val y: Double){
     override fun toString(): String {
         return "[$x, $y]"
     }
 }
 
-fun Vertex.minus(x: Float, y: Float): Vertex = Vertex(this.x-x, this.y - y)
-fun Vertex.plus(x: Float, y: Float): Vertex = Vertex(this.x + x, this.y + y)
+fun Vertex.minus(x: Double, y: Double): Vertex = Vertex(this.x-x, this.y - y)
+fun Vertex.plus(x: Double, y: Double): Vertex = Vertex(this.x + x, this.y + y)
